refactor(engine): subdivide loose modules into concern folders
I folded the multi-file engine concerns that were sitting as loose files under internal/engine/ into per-concern folders, matching the structure standard. Pure layout move with no behaviour change: network.rs + its tests become network/{mod,tests}.rs; volume.rs + volumes_list.rs become volume/{mod,list}.rs; container_fields.rs and container_security.rs move under the existing container/ as fields.rs and security.rs; and commit_export.rs + config_digests.rs become image/{export,digests}.rs behind a thin image/mod.rs. I used git mv to keep history, rewired the mod declarations and every cross-module path, and left the genuinely single-concern files flat. The build, clippy, and the full unit and non-engine integration suites pass unchanged.

Closes #553
