+++ internal/engine/image/digests.rs
